+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Satisfies the new adserver-graphql schema, adding the advertisement field into the sponsored product directly.
+
 ## [0.4.0] - 2023-09-21
 
 ### Changed